This commit has added the following to my project:
Updated the SignUpForm class to now include a
validate_email method that validates ensures a
unique email is inputted in the
templates/sign_up_page.html.

Updated the templates/sign_up_page.html and
templates/login_page.html so that there is no more
autocomplete of previous inputs in the password
and confirm password fields to prevent exposure of
data.

Implemented the LoginForm class in the forms.py
file.

Implemented the login page of the website as can
be seen in the templates/login_page.html.

Implemented a button to redirect users to either
the templates/sign_up_page.html or
templates/login_page.html if they already have an
account or have not created it respective to each
page.

Implemented the Bcrypt and LoginManager in the
__init__.py. This is important for password
encryption in the database and handling login.

Updated the User class in models.py to
have a password property as well as a property
setter (this is where the inputted password gets
encrypted into a password hash and gets assigned
to the password_hash field).

Updated the User class in models.py to now have a
method to validate a password that is submitted
against the stored password in the database by
using the instance of the Bcrypt class.

Updated the User class in models.py to now be a
child of UserMixin for future implementation.

Implemented load_user function by utilising an
instance of LoginManager and a decorator called
user_loader from LoginManager. This way, the
user session can be tracked in other pages in the
future.

Updated the login_page function in routes.py so it
now utilise the LoginForm from the forms.py. Then
carrys out form validation and matching the user
information submitted against the database.

Updated both the login_page function and
sign_up_page function in routes.py to now redirect
the user upon successfully signing up and logging
in.