This next big commit is the implementation of the Interviewer Chat Bot Page and some other important fixes and additions and removals.
New additions:

The chat_bot_page.html has been added. This also would
lead to extra implementation within existing files
such as __init__.py and routes.py as well as
index.html in accordance to the needs of the
chat_bot_page.html.

The chat_box_style.css has been added to provide
simple styling separate from bootstrap to help set
up a user friendly interactable chat bot page.

The chat_bot_updater.jss has been added and is key
to allow for the chat_bot_page.html to be updated
as the user interacts with it. This is also where
the form has been modified in chat_bot_page.html
so that it allows for the ability of sending the
user prompt to another url where the user's prompt
can interact with the processing of the info in
the python backend as well as the API. Then the
info is retrieved in chunks and added to the
chat_bot_page.html in a similar method to ChatGPT.

The error_pages folder has been added alongside
common errors that can occur when using the
website. This includes the 401_error_page.html,
404_error_page.html and 405_error_page.html as
these errors are the most common/most likely, so a
user friendly page can help users navigate through
the error.

Modifications/implementations:

__init__.py has been modified to now have
necessary imports to add the OpenAI API so it is
accessible throughout the python files.
It is important to add the API key from OpenAI has
to be added into the virtual enviroment for python
in order for it to be accessed in a safe way. This
also ensures that it is not exposed to the public
and keep it safe.

forms.py has been modified to now include a
ChatBotForm. This is to be used in the
chat_bot_page.html.

models.py has been modified to now include a login
manager that can keep track of whether the user is
logged in or not. This can be further seen in the
routes.py and how it is managed.

routes.py has been modified to now include the
route to the chat_bot_page.html. Additionally,
there is another route to complement the
chat_bot_page.html. The "/ask" route allows for
the communication with the OpenAI API and allow
for processing of the user input and returning the
stream of data back to the chat_bot_page.html.
Finally, within the routes.py, the custom pages
have been implemented as well. These all can be
found within error_pages folder.

base.html has been modified to now also include
the chat_bot_page.html as well as showing
different pages based on if the user is logged in
or not. Finally, extra jinja syntax has been used
to allow for inclusions of extra CSS files.

index.html has been modified to provide different
content based on whether the user is logged in or
not.

Side Note:

Other files may also have code that is not
commented out. These will be completed in a future
commit and the next stage of updates will be
included.

There will also be another commit which
involves cleaning up the directory such as
deleting files and general fixes to make it
overall cleaner.