Documentation/Clean-up of the project:
This is the next key commit that will include documentation of the files that contain the core logic that the website runs on. Namely, the python and javascript files will be commented heavily whilst the HTML and CSS files will be left out.

The files that have been removed were added in the initial point of the project where Tailwind CSS was going to be used. Due to difficulties and spending too much time making it compatible with the project, it has been abandoned and has long been due for removal to make everything cleaner.